tic-tac-toe/requirements.txt
Django==1.5.5
djangorestframework
